+++ modules/em/include/def.h
@@ -13,6 +13,7 @@
 
 IMPEM_BEGIN_NAMESPACE
 
+// Deprecated: just use 'double' instead
 typedef double emreal;
 const double EPS = 1e-16;