+++ modules/em/src/density_map_volumetrics.cpp
@@ -29,7 +29,7 @@ Float get_threshold_for_approximate_volume(DensityMap* d,
   Float voxelVolume = d->get_spacing() * d->get_spacing() * d->get_spacing();
   long numVoxelsNeeded = desired_volume / voxelVolume;
   long mapSizeInVoxels = d->get_number_of_voxels();
-  std::vector<emreal> data(mapSizeInVoxels);
+  std::vector<double> data(mapSizeInVoxels);
   // This costly loop could probably be replaced by a memcopy
   // provided I had access to d->data_
   for (long l = 0; l < mapSizeInVoxels; ++l) {
@@ -38,7 +38,7 @@ Float get_threshold_for_approximate_volume(DensityMap* d,
   std::sort(data.begin(), data.end());
   // If the volume is bigger than the map, just use the lowest density
   // voxel as the threshold
-  emreal threshold = data[std::max(mapSizeInVoxels - numVoxelsNeeded, 0L)];
+  double threshold = data[std::max(mapSizeInVoxels - numVoxelsNeeded, 0L)];
   return static_cast<Float>(threshold);
 }